refactor(extension): share remote mcp connect-and-persist helper
